[FEATURE] Basic visual type support
The idea here is based on oembed resource types.

This makes it so thte Visual directive can support both screenshots and
screencasts, even though the node tree will be slightly different in
each case.

The author can specify a type manually using the :type: option, but that
shouldn't be the normal way to use this.

Instead, subclasses of Visual should supply the appropriate
default_type, allowing the subclass to specify the type it expects. If
only certain types are supported, the class should override these:

- Visual.option_spec['type']
- Visual.type
- Visual.default_type

So, in the future there will be Screenshot and Screencast directives
based on the Visual directive. They will all use the visual node for
processing. Other directives, and possibly visual domains will all be
able to use this.