[FEATURE] Asset Dictionaries to track info about visuals
An asset is an image, video, or some such. It is the internal, underlying concept
for "visuals", which are the user-facing components and the API.

The idea of "Assets" might be useful for more than just the Visuals use-case.
Other possible uses include tracking images stored externally in a
DAM (Digital Asset Management) system.