[FEATURE] Monkey Patch Builder.finish()
Monkey patching seems to be the only way to get access after doc-writing in all builders.

This allows access to builder.finish_tasks *after* the write process, even if it's serial.
When it's serial, the tasks are run upon adding instead of during builder.finish().
Thus, monkey patching makes it possible to add the tasks during builder.finish().

The current task copies a placeholder image.