[FEATURE] promote assets; add asset backends; add the assets statemachine
This reorganizes, renames, and breaks up rst.processing and utils.assets.

Assets is now a top-level package, as it is more than a mere utility.
Much of what was in processing has been renamed/reorganized into:
assets.statemachine and assets.visual_asset_bridge

In part, this means that placeholders and temp images will be handled by
a specialized asset backend that fills in when no other backend can provide the asset.